+++ lumen_runtime/src/test/proptest.rs
@@ -610,6 +610,32 @@ pub fn without_number_errors_badarg(
     );
 }
 
+pub fn without_timer_reference_errors_badarg(
+    source_file: &'static str,
+    native: fn(&Process, Term) -> exception::Result<Term>,
+) {
+    run(
+        source_file,
+        |arc_process| {
+            (
+                Just(arc_process.clone()),
+                super::strategy::term::is_not_reference(arc_process.clone()),
+            )
+        },
+        |(arc_process, timer_reference)| {
+            prop_assert_badarg!(
+                native(&arc_process, timer_reference,),
+                format!(
+                    "timer_reference ({}) is not a local reference",
+                    timer_reference
+                )
+            );
+
+            Ok(())
+        },
+    );
+}
+
 pub enum FirstSecond {
     First,
     Second,
